pyproject.toml
[tool.poetry]
name = "hackrx"
version = "0.1.0"
description = ""
authors = ["Quantum-Codes <87054411+Quantum-Codes@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
requests = "^2.32.4"
pypdf = "^5.9.0"
langchain-community = "^0.3.27"
langchain-ollama = "^0.3.6"
chromadb = "^1.0.15"
google-genai = "^1.28.0"
python-dotenv = "^1.1.1"
langchain-google-genai = "^2.1.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
